Add new ClojureCompileTask
Currently this is quite hacky and ugly, but it works for now.

In particular we don't check for in-ns when checking for namespaces.
Such files cannot be easily caught by the SourceTasks filter capabilities.
I doubt there is a useful example with a namespace introduced by in-ns
but not ns.

Fixes #18